Removed current events and implemented notifications.
For me having current event buttons are a bit redundant due to that one will be able to get to current events by clicking one of the notifications, which in theory will bring you to the event page, etc. If anyone has any points bring them forth and I will try to fix or implement the request.